+++ secretmanager/api/SecretManagerTest/SamplesTest.cs
@@ -12,10 +12,10 @@
 // License for the specific language governing permissions and limitations under
 // the License.
 
-using System.IO;
-using Xunit;
 using System;
+using System.IO;
 using System.Linq;
+using Xunit;
 
 using Google.Api.Gax.ResourceNames;
 using Google.Cloud.SecretManager.V1Beta1;
@@ -42,7 +42,7 @@ namespace GoogleCloudSamples
 
         public SampleTests()
         {
-            if (s_projectId == null || s_projectId == "")
+            if (String.IsNullOrEmpty(s_projectId))
             {
                 throw new Exception("missing GOOGLE_PROJECT_ID");
             }
@@ -106,55 +106,6 @@ namespace GoogleCloudSamples
             Run("delete", s_projectId, secret1Id);
             Run("delete", s_projectId, secret2Id);
             Run("delete", s_projectId, secret3Id);
-        }
-    }
-
-    /// <summary>
-    /// Runs the quickstart and tests behavior
-    /// </summary>
-    public class QuickStartTests
-    {
-        private static readonly string s_projectId = Environment.GetEnvironmentVariable("GOOGLE_PROJECT_ID");
-
-        readonly CommandLineRunner _quickStart = new CommandLineRunner()
-        {
-            VoidMain = QuickStart.Main,
-            Command = "QuickStart"
-        };
-
-        protected ConsoleOutput Run(params string[] args)
-        {
-            return _quickStart.Run(args);
-        }
-
-        public QuickStartTests()
-        {
-            if (s_projectId == null || s_projectId == "")
-            {
-                throw new Exception("missing GOOGLE_PROJECT_ID");
-            }
-        }
-
-        [Fact]
-        public void TestRun()
-        {
-            string ts = $"{DateTime.Now.ToString("yyyyMMddHHmmssfff")}";
-            string secretId = $"csharp-quickstart-{ts}-1";
-
-            // Run quickstart
-            var output = _quickStart.Run(s_projectId, secretId);
-
-            // Verify output
-            Assert.Equal(0, output.ExitCode);
-            Assert.Contains($"Plaintext: my super secret data", output.Stdout);
-
-            // Delete the secret created by quickstart
-            SecretManagerServiceClient client = SecretManagerServiceClient.Create();
-
-            client.DeleteSecret(new DeleteSecretRequest
-            {
-                SecretName = new SecretName(s_projectId, secretId),
-            });
         }
     }
 }
